Don't unset $SHELL
The reason for unsetting it in the first place was as follows:

Originally I wrote 'repo_status' in Python using pygit2 (libgit2).

Then, because Python startup time is unacceptible in a prompt,
I rewrote repo_status in Nim.

Then, I rewrote my prompt from shell -> Zig, so I rewrote repo_status
in Zig so I could include it as a library.

To know what escapes to show if repo_status was used in a bash or zsh prompt:
* the Python version of repo_status guessed the shell from its parent process
* the Nim version looked at $SHELL (and only supported zsh)
* the Zig version requires you to tell it explicitly via cli argument

So during the time my Zig prompt was shelling out to a Nim repo_status,
I unset $SHELL so 'prompt' would work correctly when used interactively.
Now, it is no longer needed.